requirements.txt
python_dotenv
python_dateutil
hikari
hikari-arc
uvloop
gcsa
peewee